+++ src/common/CMakeLists.txt
@@ -86,6 +86,7 @@ add_library(citra_common STATIC
     logging/log.h
     logging/text_formatter.cpp
     logging/text_formatter.h
+    logging/types.h
     math_util.h
     memory_detect.cpp
     memory_detect.h
